player generation when starting a career

+++ Footballista.Game.Domain/Clubs/Teams/Rules/PlayerNumberMustBeBetween1And99.cs
@@ -4,13 +4,13 @@ namespace Footballista.Game.Domain.Clubs.Teams.Rules
 {
     internal class PlayerNumberMustBeBetween1And99 : IBusinessRule
     {
-        private readonly int _number;
+        private readonly short _number;
 
         private static readonly BoundedRange<int> _allowedNumbers = BoundedRange<int>.CreateIncluded(1, 99);
 
         public string Message => "Number must be between 1 and 99.";
 
-        public PlayerNumberMustBeBetween1And99(int number)
+        public PlayerNumberMustBeBetween1And99(short number)
         {
             _number = number;
         }
